Fix inline aliases not being loaded when extracting dev requirements, fixes #8954
