+++ src/Composer/DependencyResolver/PoolBuilder.php
@@ -88,7 +88,7 @@ class PoolBuilder
      * @param int[] $stabilityFlags an array of package name => BasePackage::STABILITY_* value
      * @psalm-param array<string, int> $stabilityFlags
      * @param array[] $rootAliases
-     * @psalm-param list<array{package: string, version: string, alias: string, alias_normalized: string}> $rootAliases
+     * @psalm-param array<string, array<string, array{alias: string, alias_normalized: string}>> $rootAliases
      * @param string[] $rootReferences an array of package name => source reference
      * @psalm-param array<string, string> $rootReferences
      */
@@ -96,7 +96,7 @@ class PoolBuilder
     {
         $this->acceptableStabilities = $acceptableStabilities;
         $this->stabilityFlags = $stabilityFlags;
-        $this->rootAliases = $this->getRootAliasesPerPackage($rootAliases);
+        $this->rootAliases = $rootAliases;
         $this->rootReferences = $rootReferences;
         $this->eventDispatcher = $eventDispatcher;
         $this->io = $io;
@@ -424,20 +424,6 @@ class PoolBuilder
 
         unset($this->skippedLoad[$name]);
         unset($this->loadedNames[$name]);
-    }
-
-    private function getRootAliasesPerPackage(array $aliases)
-    {
-        $normalizedAliases = array();
-
-        foreach ($aliases as $alias) {
-            $normalizedAliases[$alias['package']][$alias['version']] = array(
-                'alias' => $alias['alias'],
-                'alias_normalized' => $alias['alias_normalized'],
-            );
-        }
-
-        return $normalizedAliases;
     }
 }
